+++ env.sh
@@ -10,5 +10,5 @@ GOFUMPT_VERSION='v0.3.1'
 go install "mvdan.cc/gofumpt@${GOFUMPT_VERSION}"
 
 # renovate: datasource=go depName=github.com/golangci/golangci-lint
-GOLANGCI_LINT_VERSION='v1.48.0'
+GOLANGCI_LINT_VERSION='v1.49.0'
 go install "github.com/golangci/golangci-lint/cmd/golangci-lint@${GOLANGCI_LINT_VERSION}"