+++ ui/packages/shared/components/src/ProfileSelector/index.tsx
@@ -15,23 +15,16 @@ import {Query} from '@parca/parser';
 import {QueryServiceClient, ProfileTypesResponse} from '@parca/client';
 import {RpcError} from '@protobuf-ts/runtime-rpc';
 import {ProfileSelection} from '@parca/profile';
-import React, {useEffect, useMemo, useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import ProfileMetricsGraph from '../ProfileMetricsGraph';
 import MatchersInput from '../MatchersInput';
 import MergeButton from './MergeButton';
 import CompareButton from './CompareButton';
 import Card from '../Card';
-import {
-  DateTimeRangePicker,
-  DateTimeRange,
-  Select,
-  Button,
-  ButtonGroup,
-  SelectElement,
-  useGrpcMetadata,
-} from '../';
+import {DateTimeRangePicker, DateTimeRange, Button, ButtonGroup, useGrpcMetadata} from '../';
 import {CloseIcon} from '@parca/icons';
 import cx from 'classnames';
+import ProfileTypeSelector from '../ProfileTypeSelector';
 
 export interface QuerySelection {
   expression: string;
@@ -51,13 +44,6 @@ interface ProfileSelectorProps {
   profileSelection: ProfileSelection | null;
   comparing: boolean;
   onCompareProfile: () => void;
-}
-
-interface WellKnownProfiles {
-  [key: string]: {
-    name: string;
-    help: string;
-  };
 }
 
 export interface IProfileTypesResult {
@@ -86,92 +72,6 @@ export const useProfileTypes = (client: QueryServiceClient): IProfileTypesResult
   return {loading, data: result, error};
 };
 
-const wellKnownProfiles: WellKnownProfiles = {
-  'block:contentions:count:contentions:count': {
-    name: 'Block Contentions Total',
-    help: 'Stack traces that led to blocking on synchronization primitives.',
-  },
-  'block:delay:nanoseconds:contentions:count': {
-    name: 'Block Contention Time Total',
-    help: 'Time delayed stack traces caused by blocking on synchronization primitives.',
-  },
-  // Unfortunately, fgprof does not set the period type and unit.
-  'fgprof:samples:count::': {
-    name: 'Fgprof Samples Total',
-    help: 'CPU profile samples observed regardless of their current On/Off CPU scheduling status',
-  },
-  // Unfortunately, fgprof does not set the period type and unit.
-  'fgprof:time:nanoseconds::': {
-    name: 'Fgprof Samples Time Total',
-    help: 'CPU profile measured regardless of their current On/Off CPU scheduling status in nanoseconds',
-  },
-  'goroutine:goroutine:count:goroutine:count': {
-    name: 'Goroutine Created Total',
-    help: 'Stack traces that created all current goroutines.',
-  },
-  'memory:alloc_objects:count:space:bytes': {
-    name: 'Memory Allocated Objects Total',
-    help: 'A sampling of all past memory allocations by objects.',
-  },
-  'memory:alloc_space:bytes:space:bytes': {
-    name: 'Memory Allocated Bytes Total',
-    help: 'A sampling of all past memory allocations in bytes.',
-  },
-  'memory:alloc_objects:count:space:bytes:delta': {
-    name: 'Memory Allocated Objects Delta',
-    help: 'A sampling of all memory allocations during the observation by objects.',
-  },
-  'memory:alloc_space:bytes:space:bytes:delta': {
-    name: 'Memory Allocated Bytes Delta',
-    help: 'A sampling of all memory allocations during the observation in bytes.',
-  },
-  'memory:inuse_objects:count:space:bytes': {
-    name: 'Memory In-Use Objects',
-    help: 'A sampling of memory allocations of live objects by objects.',
-  },
-  'memory:inuse_space:bytes:space:bytes': {
-    name: 'Memory In-Use Bytes',
-    help: 'A sampling of memory allocations of live objects by bytes.',
-  },
-  'mutex:contentions:count:contentions:count': {
-    name: 'Mutex Contentions Total',
-    help: 'Stack traces of holders of contended mutexes.',
-  },
-  'mutex:delay:nanoseconds:contentions:count': {
-    name: 'Mutex Contention Time Total',
-    help: 'Time delayed stack traces caused by contended mutexes.',
-  },
-  'process_cpu:cpu:nanoseconds:cpu:nanoseconds:delta': {
-    name: 'Process CPU Nanoseconds',
-    help: 'CPU profile measured by the process itself in nanoseconds.',
-  },
-  'process_cpu:samples:count:cpu:nanoseconds:delta': {
-    name: 'Process CPU Samples',
-    help: 'CPU profile samples observed by the process itself.',
-  },
-  'parca_agent_cpu:samples:count:cpu:nanoseconds:delta': {
-    name: 'CPU Samples',
-    help: 'CPU profile samples observed by Parca Agent.',
-  },
-};
-
-function profileSelectElement(name: string): SelectElement {
-  const wellKnown = wellKnownProfiles[name];
-  if (wellKnown === undefined) return {active: <>{name}</>, expanded: <>{name}</>};
-
-  const title = wellKnown.name.replace(/ /g, '\u00a0');
-  return {
-    active: <>{title}</>,
-    expanded: (
-      <>
-        <span>{title}</span>
-        <br />
-        <span className="text-xs">{wellKnown.help}</span>
-      </>
-    ),
-  };
-}
-
 const ProfileSelector = ({
   queryClient,
   querySelection,
@@ -188,27 +88,6 @@ const ProfileSelector = ({
     data: profileTypesData,
     error,
   } = useProfileTypes(queryClient);
-  const profileNames = useMemo(() => {
-    return (error === undefined || error == null) &&
-      profileTypesData !== undefined &&
-      profileTypesData != null
-      ? profileTypesData.types
-          .map(
-            type =>
-              `${type.name}:${type.sampleType}:${type.sampleUnit}:${type.periodType}:${
-                type.periodUnit
-              }${type.delta ? ':delta' : ''}`
-          )
-          .sort((a: string, b: string): number => {
-            return a.localeCompare(b);
-          })
-      : [];
-  }, [profileTypesData, error]);
-
-  const profileLabels = profileNames.map(name => ({
-    key: name,
-    element: profileSelectElement(name),
-  }));
 
   const [timeRangeSelection, setTimeRangeSelection] = useState(
     DateTimeRange.fromRangeKey(querySelection.timeSelection)
@@ -295,12 +174,12 @@ const ProfileSelector = ({
               <CloseIcon />
             </button>
           )}
-          <Select
-            items={profileLabels}
+          <ProfileTypeSelector
+            profileTypesData={profileTypesData}
+            loading={profileTypesLoading}
             selectedKey={selectedProfileName}
             onSelection={setProfileName}
-            placeholder="Select profile..."
-            loading={profileTypesLoading}
+            error={error}
           />
           <MatchersInput
             queryClient={queryClient}