+++ ui/packages/shared/components/src/Select/index.tsx
@@ -91,7 +91,7 @@ const Select = ({
                 <Listbox.Options
                   className={cx(
                     width !== undefined ? `w-${width}` : '',
-                    'absolute z-10 mt-1 bg-gray-50 dark:bg-gray-900 dark:border-gray-600 shadow-lg rounded-md py-1 text-base ring-1 ring-black ring-opacity-5 overflow-auto focus:outline-none sm:text-sm'
+                    'absolute z-10 mt-1 bg-gray-50 dark:bg-gray-900 dark:border-gray-600 shadow-lg rounded-md py-1 text-base ring-1 ring-black dark:ring-white ring-opacity-5 dark:ring-opacity-20 overflow-auto focus:outline-none sm:text-sm'
                   )}
                 >
                   {loading === true ? (